z_logger/requirements.txt
requests
mysql-connector-python
flask
flask-mysql
asyncssh
scapy